refactor(CustomComboBox): move `search` from effects to class method
