+++ packages/retail-ui/components/ComboBox/__tests__/ComboBox-test.tsx
@@ -6,6 +6,10 @@ import InputLikeText from '../../internal/InputLikeText';
 import MenuItem from '../../MenuItem/MenuItem';
 import Menu from '../../Menu/Menu';
 import { delay } from 'retail-ui/lib/utils';
+import CustomComboBox, {
+  DELAY_BEFORE_SHOW_LOADER,
+  LOADER_SHOW_TIME
+} from '../../CustomComboBox/CustomComboBox';
 
 function clickOutside() {
   const event = document.createEvent('HTMLEvents');
@@ -646,6 +650,298 @@ describe('ComboBox', () => {
       it("doesn't run search if menu is open", () => {
         clickOnInput(wrapper);
         expect(getItems).toHaveBeenCalledTimes(0);
+      });
+    });
+  });
+
+  describe('Search', () => {
+    const query = 'one';
+    const items = ['one', 'two'];
+
+    // TODO check request status after merge #1168
+    // TODO test with `twice` prefix should check that items received from last request
+    it('without delay', async () => {
+      const getItems = jest.fn(() => Promise.resolve(items));
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toBeCalledWith(query);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it(`with delay < ${DELAY_BEFORE_SHOW_LOADER}`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER - 200), Promise.resolve(items)
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toBeCalledWith(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it(`with delay > ${DELAY_BEFORE_SHOW_LOADER}`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER + 200), Promise.resolve(items)
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toBeCalledWith(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: true,
+        opened: true
+      });
+
+      await delay(LOADER_SHOW_TIME);
+      await delay(0);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it('rejected without delay', async () => {
+      const getItems = jest.fn(() => Promise.reject());
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toBeCalledWith(query);
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true
+      });
+    });
+
+    it(`rejected with delay < ${DELAY_BEFORE_SHOW_LOADER}`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER - 200), Promise.reject()
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toBeCalledWith(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true
+      });
+    });
+
+    it(`rejected with delay > ${DELAY_BEFORE_SHOW_LOADER}`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER + 200), Promise.reject()
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toBeCalledWith(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: true,
+        opened: true
+      });
+
+      await delay(LOADER_SHOW_TIME);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true
+      });
+    });
+
+    it('twice without delay', async () => {
+      const getItems = jest.fn(() => Promise.resolve(items));
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+      wrapper.instance().search(query);
+
+      await delay(0);
+
+      expect(getItems).toHaveBeenCalledTimes(2);
+      expect(getItems).toBeCalledWith(query);
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it(`twice with delay < ${DELAY_BEFORE_SHOW_LOADER}`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER - 250), Promise.resolve(items)
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER - 300);
+
+      wrapper.instance().search(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER);
+
+      expect(getItems).toHaveBeenCalledTimes(2);
+      expect(getItems).toBeCalledWith(query);
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it(`twice with delay < ${DELAY_BEFORE_SHOW_LOADER} loader`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER - 100), Promise.resolve(items)
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER - 200);
+
+      wrapper.instance().search(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER - 100);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: true,
+        opened: true
+      });
+
+      await delay(LOADER_SHOW_TIME + 100);
+
+      expect(getItems).toHaveBeenCalledTimes(2);
+      expect(getItems).toBeCalledWith(query);
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it(`twice with delay > ${DELAY_BEFORE_SHOW_LOADER}`, async () => {
+      const getItems = jest.fn(
+        async () => (
+          await delay(DELAY_BEFORE_SHOW_LOADER + 200), Promise.resolve(items)
+        )
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER - 300);
+
+      wrapper.instance().search(query);
+
+      await delay(300);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: true,
+        opened: true
+      });
+
+      await delay(LOADER_SHOW_TIME + 100);
+
+      expect(getItems).toHaveBeenCalledTimes(2);
+      expect(getItems).toBeCalledWith(query);
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
+      });
+    });
+
+    it('twice with slow then fast requests', async () => {
+      const delays = [
+        DELAY_BEFORE_SHOW_LOADER + 200,
+        DELAY_BEFORE_SHOW_LOADER - 200
+      ];
+      const getItems = jest.fn(
+        async () => (await delay(delays.shift() || 0), Promise.resolve(items))
+      );
+      const wrapper = mount<ComboBox<string>>(<ComboBox getItems={getItems} />);
+
+      wrapper.instance().search(query);
+
+      await delay(300);
+
+      wrapper.instance().search(query);
+
+      await delay(DELAY_BEFORE_SHOW_LOADER - 300);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: true,
+        opened: true
+      });
+
+      await delay(200);
+
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: true,
+        opened: true
+      });
+
+      await delay(LOADER_SHOW_TIME - 200);
+
+      expect(getItems).toHaveBeenCalledTimes(2);
+      expect(getItems).toBeCalledWith(query);
+      expect(wrapper.find(CustomComboBox).instance().state).toMatchObject({
+        loading: false,
+        opened: true,
+        items
       });
     });
   });
